Overloaded [] operator, adds rows and columns functions
Added a "[ " and " ]" padding in the print function (<<).

Renames _lines to _ rows.

Added member functions that read the number of rows and columns of a
matrix.
Overloaded the [] operator to receive two size_type numbers i and j and
returns the entry [i,j] of the Matrix (starting at [0,0]).

Also fixed a few mistakes in the previous code.