Adds * between matrix and vector, adds iterators to Matrix class.
Overloaded operator * to multiply a matrix (on the left) by a vector on
the right (with the right dimensions).

Changed the implementation of the Matrix's operator << (now it is not a
friend).

Added iterators to the matrix class (they go through a whole row before
going to the next one). Also added begin and end functions to this
class.