Adds row iterators, row exchanges, multiplication by scalar and linear combinations.
Adds iterators (including const) to access a row by implementing
row_begin and row_end (and also row_cbegin and row_cend) which receive a
row number and return a (possibly const) iterator at the beginning or
right after the end of this row.

Adds row operations such as row exchange, multiplication of a row by a
double and linear combination of rows.