Changed a few details in the code (unused variables, prefix ++ instead of postfix)
Removed an unused variable in Matrix::operator* (Matrix * vector)
function.

Changed the ++ to prefix instead of postfix on (scalar * Matrix) on the
iterator to gain performance.

Removed the Matrix:: namespace in the private size() function of the
Matrix class.

Reordered the initializers on the Matrix constructor that receives a
reference to a vector of vectors of doubles to match the variables'
definition order in the Matrix class.