Corrected some casts and inline mistakes, changed _rows and _columns types to unsigned
Matrix:
- Changed the type of _rows and _columns to unsigned int (and changed
the necessary functions declarations)
- Removed the inline declaration of functions defined inside the Matrix
class; They are inlined by default. Left the comment that they are
inlined to optimize performance just to remember us;
- Inlined the operation() on Matrix;
- Named the portions of the Matrix class that defines functions that
return iterators and elementary row operations;
- Changed all casts to static_cast<TYPE>(), since they are safer (avoid
any possible unwanted reinterpret_casts).

Decomposition:
- removed unused #includes in the header file;
- It makes sense to define the function in a source file since it should
*not* be inlined. We want the lu_decomp to throw exceptions, not some
other function that uses it. Also, it is weird to inline a function this
complicated;
- Adjusted the necessary index types to unsigned, in order to avoid
unnecessary casts as it accesses the matrix's entries.